+++ src/createAnimatedComponent/JSPropUpdater.ts
@@ -3,8 +3,8 @@ import {
   NativeModules,
   findNodeHandle,
 } from 'react-native';
-import { nativeShouldBeMock } from './reanimated2/PlatformChecker';
-import type { StyleProps } from './reanimated2';
+import { nativeShouldBeMock } from '../reanimated2/PlatformChecker';
+import type { StyleProps } from '../reanimated2';
 
 interface ListenerData {
   viewTag: number;
