+++ src/createAnimatedComponent/setAndForwardRef.ts
